Package crud added in api

+++ app/Http/Controllers/Api/VehicleController.php
@@ -15,7 +15,7 @@ class VehicleController extends Controller
         if ($request->type) {
             $vechicles =  $vechicles->where('type', $request->type);
         }
-        $vechicles = $vechicles->get();
+        $vechicles = $vechicles->where('user_id', auth()->id())->get();
 
         return apiResourceResponse(VechicleResource::collection($vechicles), 'Vechicles List');
     }
@@ -43,6 +43,8 @@ class VehicleController extends Controller
 
     public function update(Request $request, Vehicle $vehicle)
     {
+        if ($vehicle->user_id != auth()->id()) return 'Permission denied';
+
         $inputs = $request->validate([
             'name' => 'nullable|string|min:2',
             'model' => 'nullable|string|min:2',
@@ -63,6 +65,8 @@ class VehicleController extends Controller
 
     public function destroy(Vehicle $vehicle)
     {
+        if ($vehicle->user_id != auth()->id()) return 'Permission denied';
+        
         detach($vehicle->image);
         $vehicle->delete();
